+++ eotsmanager/cmd/eotsd/daemon/flags.go
@@ -1,6 +1,9 @@
 package daemon
 
 const (
+	keyNameFlag           = "key-name"
+	eotsPkFlag            = "eots-pk"
+	passphraseFlag        = "passphrase"
 	forceFlag             = "force"
 	rpcListenerFlag       = "rpc-listener"
 	rpcClientFlag         = "rpc-client"