+++ CHANGELOG.md
@@ -37,9 +37,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
 * [#251](https://github.com/babylonlabs-io/finality-provider/pull/251) Add nlreturn lint
 * [#252](https://github.com/babylonlabs-io/finality-provider/pull/252) Remove interceptors and use context
 * [#266](https://github.com/babylonlabs-io/finality-provider/pull/266) Change default config
+* [#262](https://github.com/babylonlabs-io/finality-provider/pull/262) Add new command to export pop
 
 ## v0.14.3